# src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Manipulação de livros e suas dependências ",
    "version" : "v1",
    "title" : "Livraria Virtual"
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/IKokiri/Livrariav/v1",
  "schemes" : [ "https" ],
  "paths" : {
...
    "/autores" : {
      "get" : {
        "tags" : [ "autor" ],
        "summary" : "Busca autores do armazenamento",
        "operationId" : "getAutores",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Autor"
              }
            }
          },
          "400" : {
            "description" : "não encontrado"
          }
        }
      },
      "post" : {
        "tags" : [ "autor" ],
        "summary" : "Adiciona um novo autor",
        "operationId" : "addAutor",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "dados do autor que será adicionado",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Autor"
          }
        } ],
        "responses" : {
          "405" : {
            "description" : "Dados inválidos"
          }
        }
      }
    },
    "/autores/{id}" : {
      "get" : {
        "tags" : [ "autor" ],
        "summary" : "Busca autor por id",
        "description" : "Retorna um unico autor",
        "operationId" : "getAutorById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "identificador do autor de retorno",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "seccess",
            "schema" : {
              "$ref" : "#/definitions/Autor"
            }
          },
          "400" : {
            "description" : "ID inválido"
          },
          "404" : {
            "description" : "Autor não encontrado"
          }
        }
      },
      "put" : {
        "tags" : [ "autor" ],
        "summary" : "Atualização de um autor",
        "description" : "Efetua a alteração de um ou varios dados de um autor",
        "operationId" : "updateAutor",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "identificador do autor que será alterado",
          "required" : true,
          "type" : "integer"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "dados do autor que serão alterados",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Autor"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Autor inválido"
          },
          "404" : {
            "description" : "Autor não encontrado"
          }
        }
      },
      "delete" : {
        "tags" : [ "autor" ],
        "summary" : "Remoção de um autor",
        "description" : "Remove um autor",
        "operationId" : "deleteAutor",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "identificador do autor que será removido",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "400" : {
            "description" : "Autor inválido"
          },
          "404" : {
            "description" : "Autor não encontrado"
          }
        }
      }
    },
...
    "/livros/{livroId}/comentarios" : {
      "post" : {
        "tags" : [ "comentario" ],
        "summary" : "Adiciona um comentario para um livro",
        "operationId" : "addComentario",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "livroId",
          "in" : "path",
          "description" : "ID do livro",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "dados do comentrario que será adicionado",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Comentario"
          }
        } ],
        "responses" : {
          "405" : {
            "description" : "Dados inválidos"
          }
        }
      }
    },
    "/livros/{livroId}/carrinho" : {
      "post" : {
        "tags" : [ "carrinho" ],
        "summary" : "Adiciona um livro no carrinho",
        "operationId" : "addLivroCarrinho",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "livroId",
          "in" : "path",
          "description" : "ID do livro",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "dados do livro que será adicionado ao carrinho",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Livro"
          }
        } ],
        "responses" : {
          "405" : {
            "description" : "Dados inválidos"
          }
        }
      },
      "delete" : {
        "tags" : [ "carrinho" ],
        "summary" : "Remoção de um livro do carrinho",
        "description" : "Remove um livro do carrinho",
        "operationId" : "deleteLivroCarrinho",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "livroId",
          "in" : "path",
          "description" : "identificador do livro que será removido do carrinho",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "400" : {
            "description" : "Livro inválido"
          },
          "404" : {
            "description" : "livro não está no carrinho encontrado"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "Livro" : {
      "type" : "object",
      "properties" : {
        "isbn" : {
          "type" : "number"
        },
        "titulo" : {
          "type" : "string"
        },
        "autor" : {
          "$ref" : "#/definitions/Autor"
        },
        "genero" : {
          "$ref" : "#/definitions/Genero"
        },
        "data" : {
          "type" : "string",
          "format" : "date-time"
        },
        "preco" : {
          "type" : "number"
        }
      },
      "xml" : {
        "name" : "Livro"
      },
      "example" : {
        "preco" : 5.962133916683182377482808078639209270477294921875,
        "data" : "2000-01-23T04:56:07.000+00:00",
        "isbn" : 0.80082819046101150206595775671303272247314453125,
        "genero" : {
          "genero" : "genero",
          "id" : 1
        },
        "titulo" : "titulo",
        "autor" : {
          "nome" : "nome",
          "id" : 6
        }
      }
    },
    "Autor" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "nome" : {
          "type" : "string"
        }
      },
      "xml" : {
        "name" : "Autor"
      },
      "example" : {
        "nome" : "nome",
        "id" : 6
      }
    },
    "Genero" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "genero" : {
          "type" : "string"
        }
      },
      "xml" : {
        "name" : "Genero"
      },
      "example" : {
        "genero" : "genero",
        "id" : 1
      }
    },
    "Comentario" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "nome" : {
          "type" : "string"
        },
        "data" : {
          "type" : "string",
          "format" : "date-time"
        },
        "comentario" : {
          "type" : "string"
        }
      },
      "xml" : {
        "name" : "Genero"
      },
      "example" : {
        "data" : "2000-01-23T04:56:07.000+00:00",
        "nome" : "nome",
        "id" : 0,
        "comentario" : "comentario"
      }
    },
...
  }
}